Move I2cRegisterAccess to Common
